+++ test/ecto/query/inspect_test.exs
@@ -86,8 +86,8 @@ defmodule Ecto.Query.InspectTest do
 
   test "fragments" do
     value = "foobar"
-    assert i(from(x in Post, where: ~f[downcase(#{x.id}) == #{^value}])) ==
-           ~s{from p in Inspect.Post, where: ~f[downcase(p.id) == ^"foobar"]}
+    assert i(from(x in Post, where: fragment("downcase(?) == ?", x.id, ^value))) ==
+           ~s{from p in Inspect.Post, where: fragment("downcase(?) == ?", p.id, ^"foobar")}
   end
 
   test "inspect all" do