+++ test/ecto/repo_test.exs
@@ -87,7 +87,7 @@ defmodule Ecto.RepoTest do
     MockRepo.get(MyModel, 123)
 
     message = ~r"value `:atom` in `where` cannot be cast to type :integer in query"
-    assert_raise Ecto.QueryError, message, fn ->
+    assert_raise Ecto.CastError, message, fn ->
       MockRepo.get(MyModel, :atom)
     end
   end